Move the resident trade and gift regressions beside the code they describe
Half of main.rs is one inline test module: 30,329 of its 61,181 lines across
359 tests. The ceiling ledger already records why that counts -- tests spend
the same budget as production code, so an extracted system takes its tests
with it. The resident trade, gift, and delivery candidate generators moved to
autonomy.rs in an earlier slice; their regressions stayed behind.

Fourteen of them move here verbatim: no test, assertion, or fixture changed,
and the suite still reports 1,265. main.rs drops from 61,181 to 60,015 lines
without touching a line of production code.

This is the first slice of #325 and #61. The same shape is available several
times over -- production code already extracted, tests still in the root
module -- and each one is a pure move that a reviewer can read as a diff of
locations rather than behaviour.
